+++ sdk/python/kfp/compiler/compiler.py
@@ -23,7 +23,7 @@ import yaml
 from kfp.dsl import _for_loop
 
 from .. import dsl
-from ._k8s_helper import K8sHelper
+from ._k8s_helper import convert_k8s_obj_to_json, sanitize_k8s_name
 from ._op_to_template import _op_to_template
 from ._default_transformers import add_pod_env
 
@@ -639,7 +639,7 @@ class Compiler(object):
     if len(pipeline_conf.image_pull_secrets) > 0:
       image_pull_secrets = []
       for image_pull_secret in pipeline_conf.image_pull_secrets:
-        image_pull_secrets.append(K8sHelper.convert_k8s_obj_to_json(image_pull_secret))
+        image_pull_secrets.append(convert_k8s_obj_to_json(image_pull_secret))
       workflow['spec']['imagePullSecrets'] = image_pull_secrets
 
     if pipeline_conf.timeout:
@@ -684,26 +684,26 @@ class Compiler(object):
         if artifact_location and not op.artifact_location:
           op.artifact_location = artifact_location
 
-      sanitized_name = K8sHelper.sanitize_k8s_name(op.name)
+      sanitized_name = sanitize_k8s_name(op.name)
       op.name = sanitized_name
       for param in op.outputs.values():
-        param.name = K8sHelper.sanitize_k8s_name(param.name)
+        param.name = sanitize_k8s_name(param.name)
         if param.op_name:
-          param.op_name = K8sHelper.sanitize_k8s_name(param.op_name)
+          param.op_name = sanitize_k8s_name(param.op_name)
       if op.output is not None and not isinstance(op.output, dsl._container_op._MultipleOutputsError):
-        op.output.name = K8sHelper.sanitize_k8s_name(op.output.name)
-        op.output.op_name = K8sHelper.sanitize_k8s_name(op.output.op_name)
+        op.output.name = sanitize_k8s_name(op.output.name)
+        op.output.op_name = sanitize_k8s_name(op.output.op_name)
       if op.dependent_names:
-        op.dependent_names = [K8sHelper.sanitize_k8s_name(name) for name in op.dependent_names]
+        op.dependent_names = [sanitize_k8s_name(name) for name in op.dependent_names]
       if isinstance(op, dsl.ContainerOp) and op.file_outputs is not None:
         sanitized_file_outputs = {}
         for key in op.file_outputs.keys():
-          sanitized_file_outputs[K8sHelper.sanitize_k8s_name(key)] = op.file_outputs[key]
+          sanitized_file_outputs[sanitize_k8s_name(key)] = op.file_outputs[key]
         op.file_outputs = sanitized_file_outputs
       elif isinstance(op, dsl.ResourceOp) and op.attribute_outputs is not None:
         sanitized_attribute_outputs = {}
         for key in op.attribute_outputs.keys():
-          sanitized_attribute_outputs[K8sHelper.sanitize_k8s_name(key)] = \
+          sanitized_attribute_outputs[sanitize_k8s_name(key)] = \
             op.attribute_outputs[key]
         op.attribute_outputs = sanitized_attribute_outputs
       sanitized_ops[sanitized_name] = op
@@ -725,7 +725,7 @@ class Compiler(object):
     pipeline_meta = _extract_pipeline_metadata(pipeline_func)
     pipeline_meta.name = pipeline_name or pipeline_meta.name
     pipeline_meta.description = pipeline_description or pipeline_meta.description
-    pipeline_name = K8sHelper.sanitize_k8s_name(pipeline_meta.name)
+    pipeline_name = sanitize_k8s_name(pipeline_meta.name)
 
     # Need to first clear the default value of dsl.PipelineParams. Otherwise, it
     # will be resolved immediately in place when being to each component.
@@ -746,7 +746,7 @@ class Compiler(object):
         if arg_name == input.name:
           arg_type = input.type
           break
-      args_list.append(dsl.PipelineParam(K8sHelper.sanitize_k8s_name(arg_name), param_type=arg_type))
+      args_list.append(dsl.PipelineParam(sanitize_k8s_name(arg_name), param_type=arg_type))
 
     with dsl.Pipeline(pipeline_name) as dsl_pipeline:
       pipeline_func(*args_list)
@@ -759,7 +759,7 @@ class Compiler(object):
     # Fill in the default values.
     args_list_with_defaults = []
     if pipeline_meta.inputs:
-      args_list_with_defaults = [dsl.PipelineParam(K8sHelper.sanitize_k8s_name(arg_name))
+      args_list_with_defaults = [dsl.PipelineParam(sanitize_k8s_name(arg_name))
                                  for arg_name in argspec.args]
       if argspec.defaults:
         for arg, default in zip(reversed(args_list_with_defaults), reversed(argspec.defaults)):
