+++ sdk/python/kfp/containers/_container_builder.py
@@ -183,8 +183,8 @@ class ContainerBuilder(object):
                                                docker_filename=docker_filename,
                                                target_image=target_image)
       logging.info('Start a kaniko job for build.')
-      from ..compiler._k8s_helper import K8sHelper
-      k8s_helper = K8sHelper()
+      from ._k8s_job_helper import K8sJobHelper
+      k8s_helper = K8sJobHelper()
       result_pod_obj = k8s_helper.run_job(kaniko_spec, timeout)
       logging.info('Kaniko job complete.')
